Enable shim by default
